+++ service/controller/rbac/resource/namespaceauth/create.go
@@ -100,19 +100,27 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		roleBinding := writeAllRoleBindingToCustomerGroup
 		existingRoleBinding, err := r.k8sClient.RbacV1().RoleBindings(ns.Name).Get(ctx, roleBinding.Name, metav1.GetOptions{})
 		if apierrors.IsNotFound(err) {
-			r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("creating rolebinding %#q in namespace %s", roleBinding.Name, ns.Name))
+			if !pkgkey.IsProtectedNamespace(ns.Name) {
+				r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("creating rolebinding %#q in namespace %s", roleBinding.Name, ns.Name))
 
-			_, err := r.k8sClient.RbacV1().RoleBindings(ns.Name).Create(ctx, roleBinding, metav1.CreateOptions{})
-			if apierrors.IsAlreadyExists(err) {
-				// do nothing
-			} else if err != nil {
-				return microerror.Mask(err)
+				_, err := r.k8sClient.RbacV1().RoleBindings(ns.Name).Create(ctx, roleBinding, metav1.CreateOptions{})
+				if apierrors.IsAlreadyExists(err) {
+					// do nothing
+				} else if err != nil {
+					return microerror.Mask(err)
+				}
+
+				r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("rolebinding %#q in namespace %s has been created", roleBinding.Name, ns.Name))
 			}
-
-			r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("rolebinding %#q in namespace %s has been created", roleBinding.Name, ns.Name))
-
 		} else if err != nil {
 			return microerror.Mask(err)
+		} else if pkgkey.IsProtectedNamespace(ns.Name) {
+			r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("deleting rolebinding %#q in namespace %s", roleBinding.Name, ns.Name))
+			err := r.k8sClient.RbacV1().RoleBindings(ns.Name).Delete(ctx, roleBinding.Name, metav1.DeleteOptions{})
+			if err != nil {
+				return microerror.Mask(err)
+			}
+			r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("rolebinding %#q in namespace %s has been deleted", roleBinding.Name, ns.Name))
 		} else if needsUpdate(roleBinding, existingRoleBinding) {
 			r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("updating rolebinding %#q in namespace %s", roleBinding.Name, ns.Name))
 			_, err := r.k8sClient.RbacV1().RoleBindings(ns.Name).Update(ctx, roleBinding, metav1.UpdateOptions{})
@@ -155,19 +163,28 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		roleBinding := writeAllRoleBindingToAutomationSA
 		_, err = r.k8sClient.RbacV1().RoleBindings(ns.Name).Get(ctx, roleBinding.Name, metav1.GetOptions{})
 		if apierrors.IsNotFound(err) {
-			r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("creating rolebinding %#q in namespace %s", roleBinding.Name, ns.Name))
+			if !pkgkey.IsProtectedNamespace(ns.Name) {
+				r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("creating rolebinding %#q in namespace %s", roleBinding.Name, ns.Name))
 
-			_, err := r.k8sClient.RbacV1().RoleBindings(ns.Name).Create(ctx, roleBinding, metav1.CreateOptions{})
-			if apierrors.IsAlreadyExists(err) {
-				// do nothing
-			} else if err != nil {
-				return microerror.Mask(err)
+				_, err := r.k8sClient.RbacV1().RoleBindings(ns.Name).Create(ctx, roleBinding, metav1.CreateOptions{})
+				if apierrors.IsAlreadyExists(err) {
+					// do nothing
+				} else if err != nil {
+					return microerror.Mask(err)
+				}
+
+				r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("rolebinding %#q in namespace %s has been created", roleBinding.Name, ns.Name))
 			}
-
-			r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("rolebinding %#q in namespace %s has been created", roleBinding.Name, ns.Name))
-
 		} else if err != nil {
 			return microerror.Mask(err)
+		} else if pkgkey.IsProtectedNamespace(ns.Name) {
+			// delete the rolebinding
+			r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("deleting rolebinding %#q in namespace %s", roleBinding.Name, ns.Name))
+			err := r.k8sClient.RbacV1().RoleBindings(ns.Name).Delete(ctx, roleBinding.Name, metav1.DeleteOptions{})
+			if err != nil {
+				return microerror.Mask(err)
+			}
+			r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("rolebinding %#q in namespace %s has been deleted", roleBinding.Name, ns.Name))
 		}
 	}