+++ service/controller/rolebindingtemplate/resource/rolebinding/create_test.go
@@ -1,16 +1,27 @@
 package rolebinding
 
 import (
+	"context"
 	"reflect"
 	"testing"
 
+	"github.com/giantswarm/k8sclient/v7/pkg/k8sclienttest"
 	"github.com/giantswarm/k8smetadata/pkg/annotation"
 	"github.com/giantswarm/k8smetadata/pkg/label"
+	"github.com/giantswarm/micrologger/microloggertest"
+	security "github.com/giantswarm/organization-operator/api/v1alpha1"
+	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	clientgofake "k8s.io/client-go/kubernetes/fake"
+	"k8s.io/client-go/kubernetes/scheme"
+	clientfake "sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	"github.com/giantswarm/rbac-operator/api/v1alpha1"
+	pkgkey "github.com/giantswarm/rbac-operator/pkg/key"
 	"github.com/giantswarm/rbac-operator/pkg/project"
+	"github.com/giantswarm/rbac-operator/service/controller/defaultnamespace/defaultnamespacetest"
 )
 
 func TestGetRoleBindingFromTemplate(t *testing.T) {
@@ -279,7 +290,304 @@ func TestGetRoleBindingFromTemplate(t *testing.T) {
 	}
 }
 
+func TestEnsureCreated(t *testing.T) {
+	testCases := []struct {
+		Name          string
+		Template      *v1alpha1.RoleBindingTemplate
+		Organizations []string
+
+		expectedRoleBindings []*rbacv1.RoleBinding
+		expectError          bool
+	}{
+		{
+			Name: "case1: create role binding in all orgs",
+			Template: &v1alpha1.RoleBindingTemplate{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "something",
+				},
+				Spec: v1alpha1.RoleBindingTemplateSpec{
+					Template: v1alpha1.RoleBindingTemplateResource{
+						ObjectMeta: metav1.ObjectMeta{
+							Name: "something",
+						},
+						RoleRef: rbacv1.RoleRef{
+							Name: "example",
+							Kind: "ClusterRole",
+						},
+						Subjects: []rbacv1.Subject{
+							{Kind: "Group", Name: "test-group"},
+							{Kind: "ServiceAccount", Name: "test-SA"},
+						},
+					},
+				},
+			},
+			Organizations:        []string{"example", "example-2"},
+			expectedRoleBindings: []*rbacv1.RoleBinding{getTestRoleBinding(), getTestRoleBindingInNS("org-example-2")},
+		},
+		{
+			Name: "case2: create role binding in one org",
+			Template: &v1alpha1.RoleBindingTemplate{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "something",
+				},
+				Spec: v1alpha1.RoleBindingTemplateSpec{
+					Template: v1alpha1.RoleBindingTemplateResource{
+						ObjectMeta: metav1.ObjectMeta{
+							Name: "something",
+						},
+						RoleRef: rbacv1.RoleRef{
+							Name: "example",
+							Kind: "ClusterRole",
+						},
+						Subjects: []rbacv1.Subject{
+							{Kind: "Group", Name: "test-group"},
+							{Kind: "ServiceAccount", Name: "test-SA"},
+						},
+					},
+					Scopes: v1alpha1.RoleBindingTemplateScopes{
+						OrganizationSelector: v1alpha1.ScopeSelector{
+							MatchLabels: map[string]string{"name": "example"},
+						},
+					},
+				},
+			},
+			Organizations:        []string{"example", "example-2"},
+			expectedRoleBindings: []*rbacv1.RoleBinding{getTestRoleBinding()},
+		},
+		{
+			Name: "case3: no matching orgs",
+			Template: &v1alpha1.RoleBindingTemplate{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "something",
+				},
+				Spec: v1alpha1.RoleBindingTemplateSpec{
+					Template: v1alpha1.RoleBindingTemplateResource{
+						ObjectMeta: metav1.ObjectMeta{
+							Name: "something",
+						},
+						RoleRef: rbacv1.RoleRef{
+							Name: "example",
+							Kind: "ClusterRole",
+						},
+						Subjects: []rbacv1.Subject{
+							{Kind: "Group", Name: "test-group"},
+							{Kind: "ServiceAccount", Name: "test-SA"},
+						},
+					},
+					Scopes: v1alpha1.RoleBindingTemplateScopes{
+						OrganizationSelector: v1alpha1.ScopeSelector{
+							MatchLabels: map[string]string{"name": "example-3"},
+						},
+					},
+				},
+			},
+			Organizations:        []string{"example", "example-2"},
+			expectedRoleBindings: []*rbacv1.RoleBinding{},
+		},
+		{
+			Name: "case4: do not create rolebinding in org-giantswarm namespace",
+			Template: &v1alpha1.RoleBindingTemplate{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "something",
+				},
+				Spec: v1alpha1.RoleBindingTemplateSpec{
+					Template: v1alpha1.RoleBindingTemplateResource{
+						ObjectMeta: metav1.ObjectMeta{
+							Name: "something",
+						},
+						RoleRef: rbacv1.RoleRef{
+							Name: "example",
+							Kind: "ClusterRole",
+						},
+						Subjects: []rbacv1.Subject{
+							{Kind: "Group", Name: "test-group"},
+							{Kind: "ServiceAccount", Name: "test-SA"},
+						},
+					},
+				},
+			},
+			Organizations:        []string{"example", "example-2", "giantswarm"},
+			expectedRoleBindings: []*rbacv1.RoleBinding{getTestRoleBinding(), getTestRoleBindingInNS("org-example-2")},
+		},
+		{
+			Name: "case5: create rolebinding in org-giantswarm namespace",
+			Template: &v1alpha1.RoleBindingTemplate{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "something",
+				},
+				Spec: v1alpha1.RoleBindingTemplateSpec{
+					Template: v1alpha1.RoleBindingTemplateResource{
+						ObjectMeta: metav1.ObjectMeta{
+							Name: "something",
+						},
+						RoleRef: rbacv1.RoleRef{
+							Name: "example",
+							Kind: "ClusterRole",
+						},
+						Subjects: []rbacv1.Subject{
+							{Kind: "ServiceAccount", Name: "test-SA", Namespace: pkgkey.FluxNamespaceName},
+						},
+					},
+				},
+			},
+			Organizations: []string{"example", "giantswarm"},
+			expectedRoleBindings: []*rbacv1.RoleBinding{
+				{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "something",
+						Namespace: "org-giantswarm",
+						Labels: map[string]string{
+							label.ManagedBy: project.Name(),
+						},
+						Annotations: map[string]string{
+							annotation.Notes: "Generated based on RoleBindingTemplate something",
+						},
+					},
+					TypeMeta: metav1.TypeMeta{
+						Kind:       "RoleBinding",
+						APIVersion: "rbac.authorization.k8s.io/v1",
+					},
+					RoleRef: rbacv1.RoleRef{
+						Name:     "example",
+						Kind:     "ClusterRole",
+						APIGroup: "rbac.authorization.k8s.io",
+					},
+					Subjects: []rbacv1.Subject{
+						{Kind: "ServiceAccount", Name: "test-SA", Namespace: pkgkey.FluxNamespaceName},
+					},
+				},
+				{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "something",
+						Namespace: "org-example",
+						Labels: map[string]string{
+							label.ManagedBy: project.Name(),
+						},
+						Annotations: map[string]string{
+							annotation.Notes: "Generated based on RoleBindingTemplate something",
+						},
+					},
+					TypeMeta: metav1.TypeMeta{
+						Kind:       "RoleBinding",
+						APIVersion: "rbac.authorization.k8s.io/v1",
+					},
+					RoleRef: rbacv1.RoleRef{
+						Name:     "example",
+						Kind:     "ClusterRole",
+						APIGroup: "rbac.authorization.k8s.io",
+					},
+					Subjects: []rbacv1.Subject{
+						{Kind: "ServiceAccount", Name: "test-SA", Namespace: pkgkey.FluxNamespaceName},
+					},
+				},
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			objects := []runtime.Object{tc.Template}
+			namespaces := []runtime.Object{
+				&corev1.Namespace{
+					ObjectMeta: metav1.ObjectMeta{
+						Name: "example",
+					},
+				},
+			}
+			for _, org := range tc.Organizations {
+				objects = append(objects, getTestOrganization(org))
+				namespaces = append(namespaces, &corev1.Namespace{
+					ObjectMeta: metav1.ObjectMeta{
+						Name: "org-" + org,
+					},
+				})
+			}
+			var k8sClientFake *k8sclienttest.Clients
+			{
+				schemeBuilder := runtime.SchemeBuilder{
+					security.AddToScheme,
+					v1alpha1.AddToScheme,
+				}
+				if err := schemeBuilder.AddToScheme(scheme.Scheme); err != nil {
+					t.Fatal(err)
+				}
+
+				k8sClientFake = k8sclienttest.NewClients(k8sclienttest.ClientsConfig{
+					CtrlClient: clientfake.NewClientBuilder().
+						WithScheme(scheme.Scheme).
+						WithRuntimeObjects(objects...).
+						Build(),
+					K8sClient: clientgofake.NewSimpleClientset(namespaces...),
+				})
+			}
+
+			r, err := New(Config{
+				K8sClient: k8sClientFake,
+				Logger:    microloggertest.New(),
+			})
+			if err != nil {
+				t.Fatal(err)
+			}
+			ctx := context.Background()
+			err = r.EnsureCreated(ctx, tc.Template)
+			if !tc.expectError && err != nil {
+				t.Fatalf("Expected success, got error %v", err)
+			}
+			if tc.expectError && err == nil {
+				t.Fatalf("Expected error, got success")
+			}
+			roleBindingList, err := k8sClientFake.K8sClient().RbacV1().RoleBindings("").List(ctx, metav1.ListOptions{})
+			if err != nil {
+				t.Fatalf("failed to get role bindings: %s", err)
+			}
+			defaultnamespacetest.RoleBindingsShouldEqual(t, tc.expectedRoleBindings, roleBindingList.Items)
+		})
+	}
+}
+
+func getTestOrganization(name string) *security.Organization {
+	return &security.Organization{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: name,
+			Labels: map[string]string{
+				"name": name,
+			},
+		},
+		Status: security.OrganizationStatus{
+			Namespace: "org-" + name,
+		},
+	}
+}
+
 func getTestRoleBinding() *rbacv1.RoleBinding {
+	return &rbacv1.RoleBinding{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "something",
+			Namespace: "org-example",
+			Labels: map[string]string{
+				label.ManagedBy: project.Name(),
+			},
+			Annotations: map[string]string{
+				annotation.Notes: "Generated based on RoleBindingTemplate something",
+			},
+		},
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "RoleBinding",
+			APIVersion: "rbac.authorization.k8s.io/v1",
+		},
+		RoleRef: rbacv1.RoleRef{
+			Name:     "example",
+			Kind:     "ClusterRole",
+			APIGroup: "rbac.authorization.k8s.io",
+		},
+		Subjects: []rbacv1.Subject{
+			{Kind: "Group", Name: "test-group"},
+			{Kind: "ServiceAccount", Name: "test-SA", Namespace: "org-example"},
+		},
+	}
+}
+
+func getTestRoleBindingInNS(namespace string) *rbacv1.RoleBinding {
 	return &rbacv1.RoleBinding{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      "something",