+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Restrict access to `org-giantswarm` namespace
+
 ## [0.39.0] - 2023-10-26
 
 ### Added